Code refactor based on peer reviews
