+++ server/routes/documents.js
@@ -1,7 +1,10 @@
 const Documents = require('../controllers/documents');
 const Search = require('../controllers/search');
+const Auth = require('../controllers/auth');
 
 module.exports = (router) => {
+  router.use(Auth.auth);
+
   router.route('/documents')
     .post(Documents.create)
     .get(Documents.all);
