+++ server/routes/roles.js
@@ -1,6 +1,9 @@
 const Roles = require('../controllers/roles');
+const Auth = require('../controllers/auth');
 
 module.exports = (router) => {
+  router.use(Auth.auth);
+
   router.route('/roles')
     .post(Roles.create)
     .get(Roles.all);
